# [STACK-2950]: Batch member update optimization for Rack flows