## Description
Created batch tasks and added them into batch_member_update rack flows to gather batch of member related information beforehand instead of calling some tasks repeatedly for each member in a batch query using for loops.

## Jira Ticket
https://a10networks.atlassian.net/browse/STACK-3178
https://a10networks.atlassian.net/browse/STACK-3179
https://a10networks.atlassian.net/browse/STACK-3180
https://a10networks.atlassian.net/browse/STACK-3181
https://a10networks.atlassian.net/browse/STACK-3184
https://a10networks.atlassian.net/browse/STACK-3185

## Technical Approach
- Created a batch task `CountBatchMembersWithIPPortProtocol()` for getting 2 dictionaries `{member_ip_address: member_count_ip}` and `{member_ip_address: member_count_ip_port_protocol}` which contains new members and old members  for a batch update in it.
- Create a batch task `GetNatPoolEntryForBatchMembers()` for getting a dictionary `{member_subnet_id: nat_pool}` specific to old members in a batch update.
- Created a batch task `DeleteSubnetAddressAndNatPoolForBatchMembers()` for releasing subnet_address and nat_pool associated to all the member in a list of old_members in a batch update.
- Modified `ValidateSubnet()` task sothat it can work on a list of members as well in case of batch update.
- Modified `MemberCreate()` and `MemberDelete()` tasks such that `member_count_ip` and `member_count_ip_port_protocol` required in it can be fetched from the dictionaries got from the `CountBatchMembersWithIPPortProtocol()` batch tasks by looking into the dictionaries and getting the counts respective to the member which is to be deleted/created.
- Added the created batch tasks  into `get_rack_vthunder_batch_update_members_flow` before starting for loops for new_members, old_members and updated_members.
- Removed the unwanted tasks in the loops after adding the batch tasks.

## Config Changes
N/A

## Manual Testing
1. Create a flavorprofile and flavor for nat-pool

stack@openstack-3:~$ openstack loadbalancer flavorprofile show fpn
```
+---------------+------------------------------------------------------------------------------------------------------------+
| Field         | Value                                                                                                      |
+---------------+------------------------------------------------------------------------------------------------------------+
| id            | 6553359e-751d-4683-84b2-6c39997388a0                                                                       |
| name          | fpn                                                                                                        |
| provider_name | a10                                                                                                        |
| flavor_data   | {"nat-pool":{"pool-name":"pool1","start-address":"10.0.12.11","end-address":"10.0.12.12","netmask":"/24"}} |
+---------------+------------------------------------------------------------------------------------------------------------+
```

stack@openstack-3:~$ openstack loadbalancer flavor show fn
```
+-------------------+--------------------------------------+
| Field             | Value                                |
+-------------------+--------------------------------------+
| id                | bcf588e8-ecc3-444f-a31b-8d624210fed7 |
| name              | fn                                   |
| flavor_profile_id | 6553359e-751d-4683-84b2-6c39997388a0 |
| enabled           | True                                 |
| description       |                                      |
+-------------------+--------------------------------------+
```

2. Create a loadbalancer using above flavor, a listener and a pool

stack@openstack-3:~$ openstack loadbalancer create --name lb1 --vip-subnet-id public-11-subnet --flavor fn
```
+---------------------+--------------------------------------+
| Field               | Value                                |
+---------------------+--------------------------------------+
| admin_state_up      | True                                 |
| created_at          | 2021-12-01T06:49:46                  |
| description         |                                      |
| flavor_id           | bcf588e8-ecc3-444f-a31b-8d624210fed7 |
| id                  | 7ae7360f-3cdc-4bb3-83d4-362f22ed3499 |
| listeners           |                                      |
| name                | lb1                                  |
| operating_status    | OFFLINE                              |
| pools               |                                      |
| project_id          | 9ef5e94c53c940239a66dbe4a1058eee     |
| provider            | a10                                  |
| provisioning_status | PENDING_CREATE                       |
| updated_at          | None                                 |
| vip_address         | 10.0.11.182                          |
| vip_network_id      | 6df8ee71-7519-4f9d-8300-44acfa2fe325 |
| vip_port_id         | 6e0f7cd2-90b6-4a3e-92fc-b2d97fd0cd2f |
| vip_qos_policy_id   | None                                 |
| vip_subnet_id       | 184910cd-74d4-4fd5-a3b2-ebad65d4cd44 |
+---------------------+--------------------------------------+
```

stack@openstack-3:~$ openstack loadbalancer listener create --name l1 --protocol http --protocol-port 80 lb1
```
+-----------------------------+--------------------------------------+
| Field                       | Value                                |
+-----------------------------+--------------------------------------+
| admin_state_up              | True                                 |
| connection_limit            | -1                                   |
| created_at                  | 2021-12-01T06:50:12                  |
| default_pool_id             | None                                 |
| default_tls_container_ref   | None                                 |
| description                 |                                      |
| id                          | 7f81dbf3-f9dd-4ac2-a500-5d5f49f9485a |
| insert_headers              | None                                 |
| l7policies                  |                                      |
| loadbalancers               | 7ae7360f-3cdc-4bb3-83d4-362f22ed3499 |
| name                        | l1                                   |
| operating_status            | OFFLINE                              |
| project_id                  | 9ef5e94c53c940239a66dbe4a1058eee     |
| protocol                    | HTTP                                 |
| protocol_port               | 80                                   |
| provisioning_status         | PENDING_CREATE                       |
| sni_container_refs          | []                                   |
| timeout_client_data         | 50000                                |
| timeout_member_connect      | 5000                                 |
| timeout_member_data         | 50000                                |
| timeout_tcp_inspect         | 0                                    |
| updated_at                  | None                                 |
| client_ca_tls_container_ref | None                                 |
| client_authentication       | NONE                                 |
| client_crl_container_ref    | None                                 |
+-----------------------------+--------------------------------------+
```

stack@openstack-3:~$ openstack loadbalancer pool create --name p1 --protocol HTTP --lb-algorithm LEAST_CONNECTIONS --listener l1
```
+----------------------+--------------------------------------+
| Field                | Value                                |
+----------------------+--------------------------------------+
| admin_state_up       | True                                 |
| created_at           | 2021-12-01T06:50:26                  |
| description          |                                      |
| healthmonitor_id     |                                      |
| id                   | 70cfaa46-c7a6-441f-82c7-17e1052fc0d4 |
| lb_algorithm         | LEAST_CONNECTIONS                    |
| listeners            | 7f81dbf3-f9dd-4ac2-a500-5d5f49f9485a |
| loadbalancers        | 7ae7360f-3cdc-4bb3-83d4-362f22ed3499 |
| members              |                                      |
| name                 | p1                                   |
| operating_status     | OFFLINE                              |
| project_id           | 9ef5e94c53c940239a66dbe4a1058eee     |
| protocol             | HTTP                                 |
| provisioning_status  | PENDING_CREATE                       |
| session_persistence  | None                                 |
| updated_at           | None                                 |
| tls_container_ref    | None                                 |
| ca_tls_container_ref | None                                 |
| crl_container_ref    | None                                 |
| tls_enabled          | False                                |
+----------------------+--------------------------------------+
```

Result on vThunder:

```
vThunder(NOLICENSE)#show running-config
!Current configuration: 283 bytes
!Configuration last updated at 06:50:25 GMT Wed Dec 1 2021
!Configuration last saved at 06:50:28 GMT Wed Dec 1 2021
!64-bit Advanced Core OS (ACOS) version 5.2.1, build 153 (Dec-11-2020,14:16)
!
!
interface management
  ip address dhcp
!
interface ethernet 1
!
interface ethernet 2
!
vrrp-a vrid 0
  floating-ip 10.0.11.164
!
ip nat pool pool1 10.0.12.11 10.0.12.12 netmask /24
!
slb service-group 70cfaa46-c7a6-441f-82c7-17e1052fc0d4 tcp
  method least-connection
!
slb virtual-server 7ae7360f-3cdc-4bb3-83d4-362f22ed3499 10.0.11.182
  port 80 http
    name 7f81dbf3-f9dd-4ac2-a500-5d5f49f9485a
    extended-stats
    source-nat pool pool1
    service-group 70cfaa46-c7a6-441f-82c7-17e1052fc0d4
!
!
cloud-services meta-data
  enable
  provider openstack
!
end
```

3. Create 5 members using following script for batch member update:

```
resource "openstack_lb_members_v2" "members_1" {
  pool_id = "70cfaa46-c7a6-441f-82c7-17e1052fc0d4"
  member {
    address       = "10.0.12.44"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 90
    name          = "m1"
    weight        = 10
  }
  member {
    address       = "10.0.12.45"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 91
    name          = "m2"
    weight        = 15
  }
  member {
    address       = "10.0.12.46"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 92
    name          = "m3"
    weight        = 15
  }
  member {
    address       = "10.0.12.47"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 93
    name          = "m4"
    weight        = 15
  }
  member {
    address       = "10.0.12.46"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 94
    name          = "m5"
    weight        = 15
  }
}
```

stack@openstack-3:~/neha$ **time terraform apply**

Terraform used the selected providers to generate the following execution plan. Resource actions are indicated with the following symbols:
  + create

Terraform will perform the following actions:

  # openstack_lb_members_v2.members_1 will be created
  + resource "openstack_lb_members_v2" "members_1" {
      + id      = (known after apply)
      + pool_id = "70cfaa46-c7a6-441f-82c7-17e1052fc0d4"
      + region  = (known after apply)

      + member {
          + address        = "10.0.12.44"
          + admin_state_up = true
          + id             = (known after apply)
          + name           = "m1"
          + protocol_port  = 90
          + subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
          + weight         = 10
        }
      + member {
          + address        = "10.0.12.45"
          + admin_state_up = true
          + id             = (known after apply)
          + name           = "m2"
          + protocol_port  = 91
          + subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
          + weight         = 15
        }
      + member {
          + address        = "10.0.12.46"
          + admin_state_up = true
          + id             = (known after apply)
          + name           = "m3"
          + protocol_port  = 92
          + subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
          + weight         = 15
        }
      + member {
          + address        = "10.0.12.46"
          + admin_state_up = true
          + id             = (known after apply)
          + name           = "m5"
          + protocol_port  = 94
          + subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
          + weight         = 15
        }
      + member {
          + address        = "10.0.12.47"
          + admin_state_up = true
          + id             = (known after apply)
          + name           = "m4"
          + protocol_port  = 93
          + subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
          + weight         = 15
        }
    }

Plan: 1 to add, 0 to change, 0 to destroy.

Do you want to perform these actions?
  Terraform will perform the actions described above.
  Only 'yes' will be accepted to approve.

  Enter a value: yes

openstack_lb_members_v2.members_1: Creating...
openstack_lb_members_v2.members_1: Still creating... [10s elapsed]
openstack_lb_members_v2.members_1: Creation complete after 11s [id=70cfaa46-c7a6-441f-82c7-17e1052fc0d4]

Apply complete! Resources: 1 added, 0 changed, 0 destroyed.

real    0m17.510s
user    0m1.928s
sys     0m0.382s

stack@openstack-3:~/neha$ openstack loadbalancer member list p1
```
+--------------------------------------+------+----------------------------------+---------------------+------------+---------------+------------------+--------+
| id                                   | name | project_id                       | provisioning_status | address    | protocol_port | operating_status | weight |
+--------------------------------------+------+----------------------------------+---------------------+------------+---------------+------------------+--------+
| 10034af1-546e-4012-ac35-5c7e4cba21b5 | m2   | 9ef5e94c53c940239a66dbe4a1058eee | ACTIVE              | 10.0.12.45 |            91 | NO_MONITOR       |     15 |
| 20e57ab3-3d10-4eeb-b5a9-8e60dbf8c13f | m5   | 9ef5e94c53c940239a66dbe4a1058eee | ACTIVE              | 10.0.12.46 |            94 | NO_MONITOR       |     15 |
| 37d3dd11-c6a5-4fff-bd48-2b6029a527e0 | m3   | 9ef5e94c53c940239a66dbe4a1058eee | ACTIVE              | 10.0.12.46 |            92 | NO_MONITOR       |     15 |
| 7985c590-5031-46ad-b8dc-9680169a9feb | m4   | 9ef5e94c53c940239a66dbe4a1058eee | ACTIVE              | 10.0.12.47 |            93 | NO_MONITOR       |     15 |
| d7b23dab-e412-4c14-9271-591c07537fa1 | m1   | 9ef5e94c53c940239a66dbe4a1058eee | ACTIVE              | 10.0.12.44 |            90 | NO_MONITOR       |     10 |
+--------------------------------------+------+----------------------------------+---------------------+------------+---------------+------------------+--------+
```

Result on vThunder:

vThunder(NOLICENSE)#show running-config
!Current configuration: 196 bytes
!Configuration last updated at 07:02:28 GMT Wed Dec 1 2021
!Configuration last saved at 07:02:31 GMT Wed Dec 1 2021
!64-bit Advanced Core OS (ACOS) version 5.2.1, build 153 (Dec-11-2020,14:16)
!
!
interface management
  ip address dhcp
!
interface ethernet 1
!
interface ethernet 2
!
vrrp-a vrid 0
  floating-ip 10.0.11.164
  floating-ip 10.0.12.176
!
ip nat pool pool1 10.0.12.11 10.0.12.12 netmask /24
!
**slb server 9ef5e_10_0_12_44 10.0.12.44
  port 90 tcp
!
slb server 9ef5e_10_0_12_45 10.0.12.45
  port 91 tcp
!
slb server 9ef5e_10_0_12_46 10.0.12.46
  port 92 tcp
  port 94 tcp
!
slb server 9ef5e_10_0_12_47 10.0.12.47
  port 93 tcp**
!
slb service-group 70cfaa46-c7a6-441f-82c7-17e1052fc0d4 tcp
  method least-connection
  **member 9ef5e_10_0_12_44 90
  member 9ef5e_10_0_12_45 91
  member 9ef5e_10_0_12_46 92
  member 9ef5e_10_0_12_46 94
  member 9ef5e_10_0_12_47 93**
!
slb virtual-server 7ae7360f-3cdc-4bb3-83d4-362f22ed3499 10.0.11.182
  port 80 http
    name 7f81dbf3-f9dd-4ac2-a500-5d5f49f9485a
    extended-stats
    source-nat pool pool1
    service-group 70cfaa46-c7a6-441f-82c7-17e1052fc0d4
!
!
cloud-services meta-data
  enable
  provider openstack
!
end
!Current config commit point for partition 0 is 0 & config mode is classical-mode


4. Using following batch update script, delete members m1. m2 and m3, update members m4 and m5 and create new members m6, m7 and m8.

```
resource "openstack_lb_members_v2" "members_1" {
  pool_id = "70cfaa46-c7a6-441f-82c7-17e1052fc0d4"

  member {
    address       = "10.0.12.54"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 95
    name          = "m6"
    weight        = 10
  }

  member {
    address       = "10.0.12.55"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 96
    name          = "m7"
    weight        = 15
  }

  member {
    address       = "10.0.12.56"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 97
    name          = "m8"
    weight        = 15
  }

  member {
    address       = "10.0.12.47"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 93
    name          = "m4_update"
    weight        = 17
    admin_state_up = false
    admin_state_up = false
  }

  member {
    address       = "10.0.12.46"
    subnet_id     = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
    protocol_port = 94
    name          = "m5_update"
    weight        = 19
    admin_state_up = false
  }
}
```

stack@openstack-3:~/neha$ **time terraform apply**
openstack_lb_members_v2.members_1: Refreshing state... [id=70cfaa46-c7a6-441f-82c7-17e1052fc0d4]

Terraform used the selected providers to generate the following execution plan. Resource actions are indicated with the following symbols:
  ~ update in-place

Terraform will perform the following actions:

  # openstack_lb_members_v2.members_1 will be updated in-place
  ~ resource "openstack_lb_members_v2" "members_1" {
        id      = "70cfaa46-c7a6-441f-82c7-17e1052fc0d4"
        # (1 unchanged attribute hidden)

      - member {
          - address        = "10.0.12.44" -> null
          - admin_state_up = true -> null
          - backup         = false -> null
          - id             = "d7b23dab-e412-4c14-9271-591c07537fa1" -> null
          - name           = "m1" -> null
          - protocol_port  = 90 -> null
          - subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1" -> null
          - weight         = 10 -> null
        }
      - member {
          - address        = "10.0.12.45" -> null
          - admin_state_up = true -> null
          - backup         = false -> null
          - id             = "10034af1-546e-4012-ac35-5c7e4cba21b5" -> null
          - name           = "m2" -> null
          - protocol_port  = 91 -> null
          - subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1" -> null
          - weight         = 15 -> null
        }
      + member {
          + address        = "10.0.12.46"
          + admin_state_up = false
          + id             = (known after apply)
          + name           = "m5_update"
          + protocol_port  = 94
          + subnet_id      = "da29e3b4-c885-4834-b4ee-228d7d1bfac1"
          + weight         = 19
…