@ytsai-a10 2021-12-13
Besides running the batch flow parallel.
The key for the enhancement is to reduce the database query. Since I/O take more time. So, I we just move the db query from loop to another loop. It may not help.